# api.yaml
openapi: 3.0.3
info:
  title: chat server
  description: this server does provide a api for chat functionality
  version: 1.0.0
servers:
  - url: 'https:chat.schlagh.com'
components:
  securitySchemes:
    authorization:
      type: authorization
      name: authorization
      in: header
      description: the json web token
...
paths:
...
  /group/public:
    post:
      description: returns all public groups, not yet implemented
      summary: get all public groups
      operationId: getGroupsPublic
      parameters:
        - $ref: '#components/securitySchemes/authorization'
        - name: body
          in: header
          required: true
          schema:
            type: object
            properties:
              search:
                type: string
                description: this value should be searched
              limit:
                type: number
                description: how many results should be returned
              start:
                type: number
                description: where to start in the list of results
              isNotPart:
                type: boolean
                description: |
                  if true: show only chats where user is not part of
                  else: show all chats
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    description: the id of the chat
                  name:
                    type: string
                    description: the name of the chat
                  description:
                    type: string
                    description: the description of the chat
                  isPublic:
                    type: boolean
                    description: is the chat public?, true by default
        '400':
          description: bad request
        '403':
          $ref: '#components/responses/NotAllowed'
        '500':
          description: internal server error
  /group:
    put:
      description: create a new group
      summary: create a new group
      operationId: createGroup
      parameters:
        - $ref: '#components/securitySchemes/authorization'
        - name: body
          in: header
          required: true
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  name:
                    type: string
                    description: the name of the chat
                  description:
                    type: string
                    description: the description of the chat
                  isPublic:
                    type: boolean
                    description: is the chat public?
              users:
                type: array
                items:
                  type: object
                  properties:
                    uid:
                      type: integer
                      description: the id of the user
                    username:
                      type: string
                      description: the username of the user
                    isAdmin:
                      type: boolean
                      description: will the user be admin in this chat
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  chatId:
                    type: number
                    description: the id of the new groupChat
        '400':
          description: bad request
        '403':
          $ref: '#components/responses/NotAllowed'
        '500':
          description: internal server error
  /group/:gcid:
    get:
      description: get the info of the groupChat
      summary: get info of groupChat
      operationId: getGroupInfo
      parameters:
        - $ref: '#components/securitySchemes/authorization'
        - name: gcid
          in: path
          required: true
          schema:
            type: integer
            description: the id of the chat
      responses:
        '200':
          description: data
          content:
            application/json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                    description: the type of the chat
                  id:
                    type: integer
                    description: the id of the chat
                  chatName:
                    type: string
                    description: the name of the chat
                  description:
                    type: string
                    description: the description of the chat
                  public:
                    type: boolean
                    description: is the chat public?
                  memberSelf:
                    type: object
                    description: infos about the member who requested the info
                    properties:
                      isAdmin:
                        type: boolean
                        description: is the member who requested the info admin?
                  members:
                    type: array
                    description: all the other members
                    items:
                      type: object
                      properties:
                        uid:
                          type: integer
                          description: the id of the member
                        username:
                          type: string
                          description: the username of the member
                        isAdmin:
                          type: boolean
                          description: is the member admin?
                        gcmid:
                          type: integer
                          description: groupchatmemberId - the id of the groupChatMember
        '400':
          description: bad request
        '403':
          $ref: '#components/responses/NotAllowed'
        '500':
          description: internal server error
    # TODO
    delete:
      description: the group chat will be deleted - not yet implemented
      summary: delete groupChat
      operationId: deleteGroupChat
      parameters:
        - $ref: '#components/securitySchemes/authorization'
        - name: gcid
          in: path
          required: true
          schema:
            type: integer
            description: the id of the chat
      responses:
        '501':
          description: not yet implemented
  /group/:gcid/chatName:
    put:
      description: |
        change the name of the group
        requesting user has to be admin
      summary: change name of groupChat
      operationId: changeNameOfGroupChat
      parameters:
        - $ref: '#components/securitySchemes/authorization'
        - name: gcid
          in: path
          required: true
          schema:
            type: integer
            description: the id of the chat
        - name: body
          in: header
          required: true
          schema:
            type: object
            properties:
              chatName:
                type: string
                description: the new name of the chat
      responses:
        '200':
          description: successful
        '400':
          description: bad request
        '403':
          $ref: '#components/responses/NotAllowed'
        '500':
          description: internal server error
  /group/:gcid/description:
    put:
      description: |
        change the description of the group
        requesting user has to be admin
      summary: change description of groupChat
      operationId: changeDescriptionOfGroupChat
      parameters:
        - $ref: '#components/securitySchemes/authorization'
        - name: gcid
          in: path
          required: true
          schema:
            type: integer
            description: the id of the chat
        - name: body
          in: header
          required: true
          schema:
            type: object
            properties:
              description:
                type: string
                description: the new description of the chat
      responses:
        '200':
          description: successful
        '400':
          description: bad request
        '403':
          $ref: '#components/responses/NotAllowed'
        '500':
          description: internal server error
  /group/:gcid/public:
    put:
      description: |
        change the chats public state
        requesting user has to be admin
      summary: change isPublic groupChat
      operationId: changeIsPublicGroupChat
      parameters:
        - $ref: '#components/securitySchemes/authorization'
        - name: gcid
          in: path
          required: true
          schema:
            type: integer
            description: the id of the chat
        - name: body
          in: header
          required: true
          schema:
            type: object
            properties:
              isPublic:
                type: boolean
                description: the new public state
      responses:
        '200':
          description: successful
        '400':
          description: bad request
        '403':
          $ref: '#components/responses/NotAllowed'
        '500':
          description: internal server error
  /group/:gcid/member/:uid:
    put:
      description: |
        a new member gets added to the group
        requesting user has to be admin
      summary: add a member to a groupChat
      operationId: addMemberToGroupChat
      parameters:
        - $ref: '#components/securitySchemes/authorization'
        - name: gcid
          in: path
          required: true
          schema:
            type: integer
            description: the id of the chat
        - name: uid
          in: path
          required: true
          schema:
            type: integer
            description: the id of the user who should be added
      responses:
        '200':
          description: successful
        '400':
          description: bad request
        '403':
          $ref: '#components/responses/NotAllowed'
        '500':
          description: internal server error
    delete:
      description: |
        a member gets removed from the group
        requesting user has to be admin
      summary: remove a member from a groupChat
      operationId: removeMemberFromGroupChat
      parameters:
        - $ref: '#components/securitySchemes/authorization'
        - name: gcid
          in: path
          required: true
          schema:
            type: integer
            description: the id of the chat
        - name: uid
          in: path
          required: true
          schema:
            type: integer
            description: the id of the user who should be added
      responses:
        '200':
          description: successful
        '400':
          description: bad request
        '403':
          $ref: '#components/responses/NotAllowed'
        '500':
          description: internal server error
  /group/:gcid/members:
    put:
      description: |
        new members are added to the group
        requesting user has to be admin
      summary: add multiple members to a groupChat
      operationId: addMembersToGroupChat
      parameters:
        - $ref: '#components/securitySchemes/authorization'
        - name: gcid
          in: path
          required: true
          schema:
            type: integer
            description: the id of the chat
        - name: body
          in: header
          required: true
          schema:
            type: object
            properties:
              users:
                type: array
                items:
                  type: object
                  properties:
                    uid:
                      type: integer
                      description: the id of the user
                    username:
                      type: string
                      description: the username of the user
      responses:
        '200':
          description: successful
        '400':
          description: bad request
        '403':
          $ref: '#components/responses/NotAllowed'
        '500':
          description: internal server error
  /group/:gcid/member/:uid/giveAdmin:
    post:
      description: |
        admin status is added to the member
        requesting user has to be admin
      summary: give member admin status
      operationId: giveMemberAdminStatus
      parameters:
        - $ref: '#components/securitySchemes/authorization'
        - name: gcid
          in: path
          required: true
          schema:
            type: integer
            description: the id of the chat
        - name: uid
          in: path
          required: true
          schema:
            type: integer
            description: the id of the user who should be added
      responses:
        '200':
          description: successful
        '400':
          description: bad request
        '403':
          $ref: '#components/responses/NotAllowed'
        '500':
          description: internal server error
  /group/:gcid/member/:uid/removeAdmin:
    post:
      description: |
        admin status is removed from the member
        requesting user has to be admin
      summary: remove admin status from member
      operationId: removeMemberAdminStatus
      parameters:
        - $ref: '#components/securitySchemes/authorization'
        - name: gcid
          in: path
          required: true
          schema:
            type: integer
            description: the id of the chat
        - name: uid
          in: path
          required: true
          schema:
            type: integer
            description: the id of the user who should be added
      responses:
        '200':
          description: successful
        '400':
          description: bad request
        '403':
          $ref: '#components/responses/NotAllowed'
        '500':
          description: internal server error
  /group/:gcid/join:
    post:
      description: only for public chats, chat is joined
      summary: join groupChat
      operationId: joinGroupChat
      parameters:
        - $ref: '#components/securitySchemes/authorization'
        - name: gcid
          in: path
          required: true
          schema:
            type: integer
            description: the id of the chat
      responses:
        '200':
          description: successful
        '400':
          description: bad request
        '403':
          $ref: '#components/responses/NotAllowed'
        '500':
          description: internal server error
  /group/:gcid/leave:
    post:
      description: |
        the requesting user leaves the chat
        if admin, there has to be at least one admin left after this action
      summary: user takes away admin rights himself
      operationId: removeAdminSelf
      parameters:
        - $ref: '#components/securitySchemes/authorization'
        - name: gcid
          in: path
          required: true
          schema:
            type: integer
            description: the id of the chat
      responses:
        '200':
          description: successful
        '400':
          description: bad request
        '403':
          $ref: '#components/responses/NotAllowed'
        '500':
          description: internal server error
  /group/:gcid/removeAdmin:
    post:
      description: |
        the requesting user takes the admin rights away himself
        requesting user has to be admin
        there has to be at least one admin left after this action
      summary: user takes away admin rights himself
      operationId: removeAdminSelf
      parameters:
        - $ref: '#components/securitySchemes/authorization'
        - name: gcid
          in: path
          required: true
          schema:
            type: integer
            description: the id of the chat
      responses:
        '200':
          description: successful
        '400':
          description: bad request
        '403':
          $ref: '#components/responses/NotAllowed'
        '500':
          description: internal server error
...